A class for the identifiers of DateTime and ID records in spatial data.

The main use of this class and creator function is for SpatialPointsDataFrames which are used with TimeOrderedRecords for the class trip.

Note

Future versions may change significantly, this class is very basic and could probably be implemented in a better way. Specifying TOR columns by formula would be a useful addition.

Slots

TOR.columns:

2-element vector of class "character"

See also

TimeOrderedRecords, trip for creating trip objects, and trip-class for that class

Examples

showClass("TimeOrderedRecords")

#> Class "TimeOrderedRecords" [package "trip"]
#> 
#> Slots:
#>                   
#> Name:  TOR.columns
#> Class:   character
#> 
#> Known Subclasses: "trip"
#> 

tor <- new("TimeOrderedRecords", TOR.columns=c("datetime", "ID"))