Assign numeric values for Argos "class"

Assign numeric values for Argos "class" by matching the levels available to given numbers. An adjustment is made to allow sigma to be specified in kilometres, and the values returned are the approximate values for longlat degrees. It is assumed that the levels are part of an "ordered" factor from least precise to most precise.

argos.sigma(x, sigma = c(100, 80, 50, 20, 10, 4, 2), adjust = 111.12)

Arguments

x

factor of Argos location quality "classes"

sigma

numeric values (by default in kilometres)

adjust

a numeric adjustment to convert from kms to degrees

Value

Numeric values for given levels.

Details

The available levels in Argos are levels=c("Z", "B", "A", "0", "1", "2", "3").

The actual sigma values given by default are (as far as can be determined) a reasonable stab at what Argos believes.

Examples

cls <- ordered(sample(c("Z", "B", "A", "0", "1", "2", "3"), 30,
                      replace=TRUE),
               levels=c("Z", "B", "A", "0", "1", "2", "3"))
argos.sigma(cls)

#>          Z          2          1          B          Z          0          0 
#> 0.89992801 0.03599712 0.08999280 0.71994240 0.89992801 0.17998560 0.17998560 
#>          A          2          2          3          B          Z          A 
#> 0.44996400 0.03599712 0.03599712 0.01799856 0.71994240 0.89992801 0.44996400 
#>          A          B          A          Z          A          3          A 
#> 0.44996400 0.71994240 0.44996400 0.89992801 0.44996400 0.01799856 0.44996400 
#>          1          2          B          3          2          Z          0 
#> 0.08999280 0.03599712 0.71994240 0.01799856 0.03599712 0.89992801 0.17998560 
#>          1          1 
#> 0.08999280 0.08999280