As ("trip", other-classes)

Coercing trip objects to other classes.

Function to create a SpatialLinesDataFrame from a trip object, resulting in a line segment for each implicit segment along the tracks. The object stores the start and end times, duration and the ID of the segment.

# S3 method for trip
as.ppp(X, ..., fatal)

# S3 method for trip
as.psp(x, ..., from, to)

explode(x, ...)

Arguments

X

trip object.

...

reserved for future methods

fatal

Logical value, see Details of as.ppp

x

trip object

from

see as.psp for that method.

to

See as.psp.

Value

ppp object

psp object

SpatialLinesDataFrame

SpatialLinesDataFrame object with each individual line segment identified by start/end time and trip ID

Examples

## Not run: ------------------------------------
#   ## Continuing the example from '?trip-methods:
# utils::example("trip-methods", package="trip",
#            ask=FALSE, echo=FALSE)
#  as(tr, "ppp")
## ---------------------------------------------
## Not run: ------------------------------------
#  ## Continuing the example from '?trip-methods:
# utils::example("trip-methods", package="trip",
#            ask=FALSE, echo=FALSE)
#  as.psp.trip(tr)
## ---------------------------------------------
## Continuing the example from '?trip-methods:
utils::example("trip-methods", package="trip",
           ask=FALSE, echo=FALSE)

Warning message:
no help found for ‘trip-methods’

spldf <- explode(tr)
Error in getTORnames(x): object 'tr' not found
summary(tr)
Error in summary(tr): object 'tr' not found