SeaWiFS ocean colour colours

Generate ocean colour colours, using the SeaWiFS scheme

oc.theme(x = 50)

oc.colors(n)

Arguments

x
Number of colours to generate as part of a theme

n
Number of colours to generate

Value

A set of colours or a theme object.

Details

This is a high-contrast palette, log-scaled originally for ocean chlorophyll.

See also

Examples

if (FALSE) {
 oc.colors(10)
 library(lattice)
 trellis.par.set(oc.theme())
 d <- data.frame(x=1:10, y=rnorm(10), tms=Sys.time() + 1:10, id=gl(2, 5))
sp::coordinates(d) <- ~x+y
## this avoids complaints later, but these are not real track data (!)
sp::proj4string(d) <- sp::CRS("+proj=laea +ellps=sphere", doCheckCRSArgs = FALSE)
tr <- trip(d, c("tms", "id"))

 tg <- tripGrid(tr)
 spplot(tg)
}