A class for sets of animal trips (track data).

An extension of SpatialPointsDataFrame by including "TimeOrderedRecords". The records within the data frame are explicitly ordered by DateTime data within IDs.

Objects from the Class

Objects can be created by calls of the form trip(obj="SpatialPointsDataFrame", TORnames="TimeOrderedRecords"). The object contains all the slots present within a SpatialPointsDataFrame, particularly data which contains columns of at least those specified by TOR.columns.

See also

trip for examples of directly using the class.

trip-accessors describes methods for accessing information on trip objects.

Examples

showClass("trip")

#> Class "trip" [package "trip"]
#> 
#> Slots:
#>                                                                               
#> Name:  TOR.columns        data  coords.nrs      coords        bbox proj4string
#> Class:   character  data.frame     numeric      matrix      matrix         CRS
#> 
#> Extends: 
#> Class "TimeOrderedRecords", directly
#> Class "SpatialPointsDataFrame", directly
#> Class "SpatialPoints", by class "SpatialPointsDataFrame", distance 2
#> Class "Spatial", by class "SpatialPointsDataFrame", distance 3
#> Class "SpatialVector", by class "SpatialPointsDataFrame", distance 3

## Examples of general methods
## Continuing the example from '?trip-methods:
utils::example("trip-methods", package="trip",
               ask=FALSE, echo=FALSE)

#> Warning: no help found for ‘trip-methods’

summary(tr)

#> Error in summary(tr): object 'tr' not found

plot(tr)

#> Error in plot(tr): object 'tr' not found

lines(tr)

#> Error in lines(tr): object 'tr' not found

dim(tr)

#> Error in eval(expr, envir, enclos): object 'tr' not found

names(tr)

#> Error in eval(expr, envir, enclos): object 'tr' not found

subset(tr, id == "2")

#> Error in subset(tr, id == "2"): object 'tr' not found

as.data.frame(tr)

#> Error in as.data.frame(tr): object 'tr' not found

tr[1:3, ]

#> Error in eval(expr, envir, enclos): object 'tr' not found

tr[, 1]

#> Error in eval(expr, envir, enclos): object 'tr' not found

tr[[1]]

#> Error in eval(expr, envir, enclos): object 'tr' not found

if (exists("porpoise")) {
    dim(porpoise)
    names(porpoise)
    porpoise[porpoise[["id"]] == "GUS", ]
}