As ("trip", other-classes)

Coercing trip objects to other classes.

Function to create a SpatialLinesDataFrame from a trip object, resulting in a line segment for each implicit segment along the tracks. The object stores the start and end times, duration and the ID of the segment.

# S3 method for trip
as.ppp(X, ..., fatal)

# S3 method for trip
as.psp(x, ..., from, to)

explode(x, ...)

Arguments

X
trip object.

...
reserved for future methods

fatal
Logical value, see Details of as.ppp

x
trip object

from
see as.psp for that method.

to
See as.psp.

Value

ppp object

psp object

SpatialLinesDataFrame

SpatialLinesDataFrame object with each individual line segment identified by start/end time and trip ID

Examples

# NOT RUN {
 d <- data.frame(x=1:10, y=rnorm(10), tms=Sys.time() + 1:10, id=gl(2, 5))
sp::coordinates(d) <- ~x+y
## this avoids complaints later, but these are not real track data (!)
sp::proj4string(d) <- sp::CRS("+proj=laea +ellps=sphere")
tr <- trip(d, c("tms", "id"))

 as(tr, "ppp")
# }# NOT RUN {
 d <- data.frame(x=1:10, y=rnorm(10), tms=Sys.time() + 1:10, id=gl(2, 5))
sp::coordinates(d) <- ~x+y
## this avoids complaints later, but these are not real track data (!)
sp::proj4string(d) <- sp::CRS("+proj=laea +ellps=sphere")
tr <- trip(d, c("tms", "id"))

 as.psp.trip(tr)
# }
 d <- data.frame(x=1:10, y=rnorm(10), tms=Sys.time() + 1:10, id=gl(2, 5))
sp::coordinates(d) <- ~x+y
## this avoids complaints later, but these are not real track data (!)
sp::proj4string(d) <- sp::CRS("+proj=laea +ellps=sphere")
tr <- trip(d, c("tms", "id"))

spldf <- explode(tr)
summary(tr)

#> 
#> Object of class trip
#>   tripID ("id") No.Records   startTime ("tms")     endTime ("tms") tripDuration
#> 1             1          5 2019-04-11 02:41:25 2019-04-11 02:41:29       4 secs
#> 2             2          5 2019-04-11 02:41:30 2019-04-11 02:41:34       4 secs
#>   tripDistance meanSpeed maxSpeed
#> 1     5.708647  5137.783 6361.983
#> 2     6.748403  6073.563 8186.395
#> 
#> Total trip duration: 8 seconds (0 hours, 8 seconds)
#> 
#> Derived from Spatial data:
#> 
#> Object of class SpatialPointsDataFrame
#> Coordinates:
#>         min       max
#> x  1.000000 10.000000
#> y -1.863011  1.888505
#> Is projected: TRUE 
#> proj4string : [+proj=laea +ellps=sphere]
#> Number of points: 10
#> Data attributes:
#>       tms                      id   
#>  Min.   :2019-04-11 02:41:25   1:5  
#>  1st Qu.:2019-04-11 02:41:27   2:5  
#>  Median :2019-04-11 02:41:29        
#>  Mean   :2019-04-11 02:41:29        
#>  3rd Qu.:2019-04-11 02:41:32        
#>  Max.   :2019-04-11 02:41:34        
#>